Remove some old type casts

+++ src/ccstruct/blobbox.h
@@ -2,7 +2,6 @@
  * File:        blobbox.h  (Formerly blobnbox.h)
  * Description: Code for the textord blob class.
  * Author:      Ray Smith
- * Created:     Thu Jul 30 09:08:51 BST 1992
  *
  * (C) Copyright 1992, Hewlett-Packard Ltd.
  ** Licensed under the Apache License, Version 2.0 (the "License");
@@ -21,7 +20,7 @@
 #define BLOBBOX_H
 
 #include <cinttypes>           // for PRId32
-#include <cmath>               // for sqrt
+#include <cmath>               // for std::sqrt
 #include <cstdint>             // for int16_t, int32_t
 #include "elst.h"              // for ELIST_ITERATOR, ELISTIZEH, ELIST_LINK
 #include "elst2.h"             // for ELIST2_ITERATOR, ELIST2IZEH, ELIST2_LINK
@@ -627,9 +626,8 @@ class TO_ROW: public ELIST2_LINK
                            float new_error) {
       para_c = new_c;
       para_error = new_error;
-      credibility =
-        (float) (blobs.length () - kErrorWeight * new_error);
-      y_origin = (float) (new_c / sqrt (1 + gradient * gradient));
+      credibility = blobs.length() -  kErrorWeight * new_error;
+      y_origin = new_c / std::sqrt(1 + gradient * gradient);
       //real intercept
     }
     void set_limits(                  //set min,max
